all: move examples to their own example_test file
